# backend/openapi/v1.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "CPX Secreto API",
        "version": "v1"
    },
    "paths": {
...
        "/api/v1/groups/{group}/editions/{edition}/draw-constraints": {
            "get": {
                "tags": [
                    "Draw constraints"
                ],
                "operationId": "listDrawConstraints",
                "parameters": [
                    {
                        "name": "group",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "edition",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Private draw constraints.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DrawConstraintCollection"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Authentication is required.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "An active administrator is required.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "The group or edition is not visible.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Draw constraints"
                ],
                "operationId": "createDrawConstraint",
                "parameters": [
                    {
                        "name": "group",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "edition",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateDrawConstraintRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Constraint created.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DrawConstraint"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Authentication is required.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "An active administrator is required.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "The group, edition, or participant is not visible.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "The constraint conflicts or makes the draw impossible.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "The request payload is invalid.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/groups/{group}/editions/{edition}/draw-constraints/copy-from-previous": {
            "post": {
                "tags": [
                    "Draw constraints"
                ],
                "operationId": "copyDrawConstraintsFromPreviousEdition",
                "parameters": [
                    {
                        "name": "group",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "edition",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Recurring exclusions copied from the previous edition.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CopiedDrawConstraints"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Authentication is required.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "An active administrator is required.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "The group or edition is not visible.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "The edition can no longer be changed.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/groups/{group}/editions/{edition}/draw-constraints/{drawConstraint}": {
            "delete": {
                "tags": [
                    "Draw constraints"
                ],
                "operationId": "deleteDrawConstraint",
                "parameters": [
                    {
                        "name": "group",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "edition",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "drawConstraint",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Constraint deleted."
                    },
                    "401": {
                        "description": "Authentication is required.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "An active administrator is required.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "The constraint is not visible.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "The edition can no longer be changed.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CopiedDrawConstraints": {
                "required": [
                    "sourceEditionId",
                    "copiedCount",
                    "skippedMissingParticipants",
                    "skippedDuplicates",
                    "skippedConflicts",
                    "data"
                ],
                "properties": {
                    "sourceEditionId": {
                        "oneOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "copiedCount": {
                        "type": "integer",
                        "minimum": 0
                    },
                    "skippedMissingParticipants": {
                        "type": "integer",
                        "minimum": 0
                    },
                    "skippedDuplicates": {
                        "type": "integer",
                        "minimum": 0
                    },
                    "skippedConflicts": {
                        "type": "integer",
                        "minimum": 0
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/DrawConstraint"
                        }
                    }
                },
                "type": "object"
            },
            "CreateDrawConstraintRequest": {
                "required": [
                    "type",
                    "giverParticipantId",
                    "receiverParticipantId"
                ],
                "properties": {
                    "type": {
                        "type": "string",
                        "enum": [
                            "must_not_pair",
                            "must_pair"
                        ]
                    },
                    "giverParticipantId": {
                        "type": "integer",
                        "minimum": 1
                    },
                    "receiverParticipantId": {
                        "type": "integer",
                        "minimum": 1
                    }
                },
                "type": "object"
            },
            "DrawConstraintCollection": {
                "required": [
                    "data"
                ],
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/DrawConstraint"
                        }
                    }
                },
                "type": "object"
            },
            "DrawConstraint": {
                "required": [
                    "id",
                    "editionId",
                    "type",
                    "giverParticipantId",
                    "receiverParticipantId",
                    "source"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "editionId": {
                        "type": "integer",
                        "example": 1
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "must_not_pair",
                            "must_pair"
                        ]
                    },
                    "giverParticipantId": {
                        "type": "integer",
                        "example": 1
                    },
                    "receiverParticipantId": {
                        "type": "integer",
                        "example": 2
                    },
                    "source": {
                        "type": "string",
                        "enum": [
                            "admin",
                            "purchase"
                        ]
                    }
                },
                "type": "object"
            },
...
            "Error": {
                "required": [
                    "message"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Não autenticado."
                    },
                    "errors": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}